+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 
 ### Changed
 - Bump `gl` to v6 and move to Node.js 18 in CI.
-- Bump `deck.gl` to v8.9.30 and `luma.gl` to v8.5.21.
+- Bump `deck.gl` to v8.8.27 and `luma.gl` to v8.5.21.
+- Narrow required interface for `ZarrPixelSource`.
 
 ## 0.13.8